\name{prng542-package}
\alias{prng542-package}
\alias{prng542}
\docType{package}
\title{
  A few PRNGs and CDF approximations
}
\description{
  In the STAT 542 lectures, it was mentioned that
    the generation of random numbers would be an interesting exercise,
    although tangentially related to the syllabus of the class. This package
    implements a few well-known pseudo-random number generators along with
    some approximation methods to generate uniform and some non-uniform
    distributions.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Gautham V

Maintainer: Gautham V <ahgamut@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ PRNG }
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
